Make mojo_device_aten_ops.py a registration list only
The file grew proportionally to the number of supported ops because the 38
ops that need real Python code lived in it alongside the 151 one-line fast
bindings. The implementations move to a new `mojo_device/aten_ops/` package,
grouped by family, and the registration list becomes one alphabetical entry
per aten name: 1370 lines -> 381.

Pure code movement: every moved function body is AST-identical to before, and
the registered op set is byte-identical (205 ops, 205 call counters, no
duplicates) both with and without test instrumentation. Registration order no
longer follows the file layout, which is safe because each aten name is
registered exactly once and `register_mojo_devices()` only iterates the
registry.

One behavior change falls out of the move and is worth knowing about.
beartype's claw hook installs `@beartype` as the *outermost* decorator, so a
function written as `@register_aten_op(...) def ...` registered the inner
`counted` wrapper with torch and never type-checked the dispatch path -- the
hints on all 38 implementations were dead under the test suite. Registering
them from the list instead puts beartype back in the call path, which
surfaced exactly one wrong hint: `aten::add_.Tensor`'s `alpha` is a Scalar,
not a float, so `x.add_(3.0, alpha=-2)` arrives with an int. Widened, with a
comment on why it differs from the coerced `float` schema args nearby.

The docs that tell contributors where to add an op (AGENTS.md step 7,
docs/CONTRIBUTING.md) now describe the two-place layout, and the few
`mojo_device_aten_ops.py:<line>` references in docs/optimization_backlog.md
that the split would have pointed past end-of-file became symbolic.

Closes #381
