+++ torch_mojo_backend/mojo_device/aten_ops/__init__.py
@@ -1,0 +1,6 @@
+"""Eager-mode aten implementations, grouped by family.
+
+`mojo_device_aten_ops.py` imports the functions defined here and binds each
+one to its aten name; nothing in this package registers anything itself, so
+the registration list stays in one readable place.
+"""
